+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/hello_world/hello_world_extension.py
@@ -9,21 +9,39 @@
 
 import os
 
-from isaacsim.examples.interactive.base_sample import BaseSampleExtension
+import omni.ext
+from isaacsim.examples.browser import get_instance as get_browser_instance
+from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
 from isaacsim.examples.interactive.hello_world import HelloWorld
 
 
-class HelloWorldExtension(BaseSampleExtension):
+class HelloWorldExtension(omni.ext.IExt):
     def on_startup(self, ext_id: str):
-        super().on_startup(ext_id)
-        super().start_extension(
-            menu_name="",
-            submenu_name="",
-            name="Hello World",
-            title="Hello World Example",
-            doc_link="https://docs.omniverse.nvidia.com/isaacsim/latest/core_api_tutorials/tutorial_core_hello_world.html",
-            overview="This Example introduces the user on how to do cool stuff with Isaac Sim through scripting in asynchronous mode.",
-            file_path=os.path.abspath(__file__),
-            sample=HelloWorld(),
+        self.example_name = "Hello World"
+        self.category = "General"
+
+        ui_kwargs = {
+            "ext_id": ext_id,
+            "file_path": os.path.abspath(__file__),
+            "title": "Hello World Example",
+            "doc_link": "https://docs.omniverse.nvidia.com/isaacsim/latest/core_api_tutorials/tutorial_core_hello_world.html",
+            "overview": "This Example introduces the user on how to do cool stuff with Isaac Sim through scripting in asynchronous mode.",
+            "sample": HelloWorld(),
+        }
+
+        ui_handle = BaseSampleUITemplate(**ui_kwargs)
+
+        # register the example with examples browser
+        get_browser_instance().register_example(
+            name=self.example_name,
+            execute_entrypoint=ui_handle.build_window,
+            ui_hook=ui_handle.build_ui,
+            category=self.category,
         )
+
+        return
+
+    def on_shutdown(self):
+        get_browser_instance().deregister_example(name=self.example_name, category=self.category)
+
         return