+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/humanoid/humanoid_example_extension.py
@@ -9,13 +9,17 @@
 
 import os
 
-from isaacsim.examples.interactive.base_sample import BaseSampleExtension
+import omni.ext
+from isaacsim.examples.browser import get_instance as get_browser_instance
+from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
 from isaacsim.examples.interactive.humanoid import HumanoidExample
 
 
-class HumanoidExampleExtension(BaseSampleExtension):
+class HumanoidExampleExtension(omni.ext.IExt):
     def on_startup(self, ext_id: str):
-        super().on_startup(ext_id)
+
+        self.example_name = "Humanoid"
+        self.category = "Policy"
 
         overview = "This Example shows an Unitree H1 running a flat terrain policy trained in Isaac Lab"
         overview += "\n\tKeybord Input:"
@@ -24,13 +28,28 @@ class HumanoidExampleExtension(BaseSampleExtension):
         overview += "\n\t\tright arrow / numpad 6: Spin Clockwise"
         overview += "\n\nPress the 'Open in IDE' button to view the source code."
 
-        super().start_extension(
-            menu_name="",
-            submenu_name="",
-            name="Humanoid",
-            title="Unitree H1 Example",
-            doc_link="https://docs.omniverse.nvidia.com/isaacsim/latest/isaac_lab_tutorials/tutorial_policy_deployment.html",
-            overview=overview,
-            file_path=os.path.abspath(__file__),
-            sample=HumanoidExample(),
+        ui_kwargs = {
+            "ext_id": ext_id,
+            "file_path": os.path.abspath(__file__),
+            "title": "Humanoid: Unitree H1",
+            "doc_link": "https://docs.omniverse.nvidia.com/isaacsim/latest/isaac_lab_tutorials/tutorial_policy_deployment.html",
+            "overview": overview,
+            "sample": HumanoidExample(),
+        }
+
+        ui_handle = BaseSampleUITemplate(**ui_kwargs)
+
+        # register the example with examples browser
+        get_browser_instance().register_example(
+            name=self.example_name,
+            execute_entrypoint=ui_handle.build_window,
+            ui_hook=ui_handle.build_ui,
+            category=self.category,
         )
+
+        return
+
+    def on_shutdown(self):
+        get_browser_instance().deregister_example(name=self.example_name, category=self.category)
+
+        return