+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/quadruped/quadruped_example_extension.py
@@ -9,13 +9,16 @@
 
 import os
 
-from isaacsim.examples.interactive.base_sample import BaseSampleExtension
+import omni.ext
+from isaacsim.examples.browser import get_instance as get_browser_instance
+from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
 from isaacsim.examples.interactive.quadruped import QuadrupedExample
 
 
-class QuadrupedExampleExtension(BaseSampleExtension):
+class QuadrupedExampleExtension(omni.ext.IExt):
     def on_startup(self, ext_id: str):
-        super().on_startup(ext_id)
+        self.example_name = "Quadruped"
+        self.category = "Policy"
 
         overview = "This Example shows an Boston Dynamics Spot running a flat terrain policy trained in Isaac Lab"
         overview += "\n\tKeybord Input:"
@@ -28,14 +31,28 @@ class QuadrupedExampleExtension(BaseSampleExtension):
 
         overview += "\n\nPress the 'Open in IDE' button to view the source code."
 
-        super().start_extension(
-            menu_name="",
-            submenu_name="",
-            name="Quadruped",
-            title="Boston Dynamic Spot Quadruped Example",
-            doc_link="https://docs.omniverse.nvidia.com/isaacsim/latest/isaac_lab_tutorials/tutorial_policy_deployment.html",
-            overview=overview,
-            file_path=os.path.abspath(__file__),
-            sample=QuadrupedExample(),
+        ui_kwargs = {
+            "ext_id": ext_id,
+            "file_path": os.path.abspath(__file__),
+            "title": "Quadruped: Boston Dynamics Spot",
+            "doc_link": "https://docs.omniverse.nvidia.com/isaacsim/latest/isaac_lab_tutorials/tutorial_policy_deployment.html",
+            "overview": overview,
+            "sample": QuadrupedExample(),
+        }
+
+        ui_handle = BaseSampleUITemplate(**ui_kwargs)
+
+        # register the example with examples browser
+        get_browser_instance().register_example(
+            name=self.example_name,
+            execute_entrypoint=ui_handle.build_window,
+            ui_hook=ui_handle.build_ui,
+            category=self.category,
         )
+
+        return
+
+    def on_shutdown(self):
+        get_browser_instance().deregister_example(name=self.example_name, category=self.category)
+
         return