+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/simple_stack/simple_stack_extension.py
@@ -10,33 +10,51 @@
 import asyncio
 import os
 
+import omni.ext
 import omni.ui as ui
-from isaacsim.examples.interactive.base_sample import BaseSampleExtension
+from isaacsim.examples.browser import get_instance as get_browser_instance
+from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
 from isaacsim.examples.interactive.simple_stack import SimpleStack
 from isaacsim.gui.components.ui_utils import btn_builder
 
 
-class SimpleStackExtension(BaseSampleExtension):
+class SimpleStackExtension(omni.ext.IExt):
     def on_startup(self, ext_id: str):
-        super().on_startup(ext_id)
-        super().start_extension(
-            menu_name="Manipulation",
-            submenu_name="",
-            name="Simple Stack",
-            title="Stack Two Cubes",
-            doc_link="https://docs.omniverse.nvidia.com/isaacsim/latest/core_api_tutorials/tutorial_core_adding_manipulator.html",
-            overview="This Example shows how to stack two cubes using Franka robot in Isaac Sim.\n\nPress the 'Open in IDE' button to view the source code.",
-            sample=SimpleStack(),
-            file_path=os.path.abspath(__file__),
+        self.example_name = "Simple Stack"
+        self.category = "Manipulation"
+
+        ui_kwargs = {
+            "ext_id": ext_id,
+            "file_path": os.path.abspath(__file__),
+            "title": "Simple Stack",
+            "doc_link": "https://docs.omniverse.nvidia.com/isaacsim/latest/core_api_tutorials/tutorial_core_adding_manipulator.html",
+            "overview": "This Example shows how to stack two cubes using Franka robot in Isaac Sim.\n\nPress the 'Open in IDE' button to view the source code.",
+            "sample": SimpleStack(),
+        }
+
+        ui_handle = SimpleStackUI(**ui_kwargs)
+
+        get_browser_instance().register_example(
+            name=self.example_name,
+            execute_entrypoint=ui_handle.build_window,
+            ui_hook=ui_handle.build_ui,
+            category=self.category,
         )
 
         return
 
-    def build_ui(self):
-        extra_stacks = self.build_default_frame()
-        self.build_extra_frames(extra_stacks)
+    def on_shutdown(self):
+        get_browser_instance().deregister_example(name=self.example_name, category=self.category)
+        return
 
-    def build_extra_frames(self, extra_stacks):
+
+class SimpleStackUI(BaseSampleUITemplate):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def build_extra_frames(self):
+
+        extra_stacks = self.get_extra_frames_handle()
         self.task_ui_elements = {}
 
         with extra_stacks: